+++ pkgs/dart_mcp_server/lib/src/utils/sdk.dart
@@ -25,15 +25,24 @@ class Sdk {
 
   /// Creates an [Sdk] from the path to the Dart SDK.
   ///
-  /// If no [dartSdkPath] is given, this will attempt to find one using
-  /// [Platform.resolvedExecutable], assuming that is the `dart` binary
-  /// under the `bin` dir of a Dart SDK.
+  /// If no [dartSdkPath] is given, this will attempt to find one by
+  /// first looking for a `DART_ROOT` environment variable, and then using
+  /// [Platform.resolvedExecutable] (assuming that is the `dart` binary
+  /// under the `bin` dir of a Dart SDK).
   ///
   /// Validates that the path is valid by checking for the `version` file.
   ///
   /// If no [flutterSdkPath] is given, this will search up from the resolved
   /// Dart SDK path to see if it is nested inside a Flutter SDK.
-  factory Sdk.find({String? dartSdkPath, String? flutterSdkPath}) {
+  factory Sdk.find({
+    String? dartSdkPath,
+    String? flutterSdkPath,
+    Map<String, String>? environment,
+  }) {
+    environment ??= Platform.environment;
+    dartSdkPath ??= environment['DART_ROOT'] ?? environment['DART_SDK'];
+    flutterSdkPath ??=
+        environment['FLUTTER_ROOT'] ?? environment['FLUTTER_SDK'];
     // Assume that we are running from the Dart SDK bin dir if not given any
     // other configuration.
     dartSdkPath ??= p.dirname(p.dirname(Platform.resolvedExecutable));
@@ -67,7 +76,7 @@ class Sdk {
           ?.child('bin')
           .child('dart${Platform.isWindows ? '.exe' : ''}') ??
       (throw ArgumentError(
-        'Dart SDK location unknown, try setting the DART_SDK environment '
+        'Dart SDK location unknown, try setting the DART_ROOT environment '
         'variable.',
       ));